Use rinetd to redirect ports
The devel config for Thentos is restrictive in terms of addresses
listened on.  Requests must (appear to) come from localhost.  A normal
Vagrant config can't do this, so we use rinetd to redirect ports on the
guest system.  Effectively, ports 7001 and 7002 on the host are mapped
to the same ports on the guest's localhost.